check whether final files exist to move out from workdir
Reviewed - check whether final files exist to move out from workdir

+++ dna_pipeline.py
@@ -12,7 +12,6 @@ Multiple options are available. To see them type --help
 @author: Jose Fernandez Navarro <jc.fernandez.navarro@gmail.com>
 """
 from hlapipeline.common import *
-import shutil
 from argparse import ArgumentParser, RawDescriptionHelpFormatter
 import os
 import sys
@@ -122,11 +121,11 @@ def main(R1_NORMAL,
         exec_command(cmd)
 
         # BamQC
-        cmd = '{} -bam sample2_final.bam --genome-gc-distr HUMAN -nt {} ' \
+        cmd = '{} -bam sample2_final.bam --genome-gc-distr HUMAN -nt {} --java-mem-size=16G ' \
               '-outdir bamQC_Normal -outformat HTML'.format(BAMQC, THREADS)
         p1 = exec_command(cmd, detach=True)
 
-        cmd = '{} -bam sample1_final.bam --genome-gc-distr HUMAN -nt {} ' \
+        cmd = '{} -bam sample1_final.bam --genome-gc-distr HUMAN -nt {} --java-mem-size=16G ' \
               '-outdir bamQC_Tumor -outformat HTML'.format(BAMQC, THREADS)
         p2 = exec_command(cmd, detach=True)
 
@@ -234,21 +233,30 @@ def main(R1_NORMAL,
         annotate_variants('combined_calls.vcf', 'annotated', ANNOVAR_DB, ANNOVAR_VERSION, THREADS)
 
         # Moving result files to output
-        shutil.move('combined_calls.vcf', '../combined_calls.vcf')
-        shutil.move('annotated.{}_multianno.vcf'.format(ANNOVAR_VERSION),
-                    '../annotated.{}_multianno.vcf'.format(ANNOVAR_VERSION))
-        shutil.move('Tumor_{}_hla_genotype_result.tsv'.format(SAMPLEID),
-                    '../Tumor_hla_genotype.tsv')
-        shutil.move('Normal_{}_hla_genotype_result.tsv'.format(SAMPLEID),
-                    '../Normal_hla_genotype.tsv')
-        shutil.move('sample1_final.bam', '../tumor_final.bam')
-        shutil.move('sample2_final.bam', '../normal_final.bam')
+
+	if os.path.isfile('combined_calls.vcf'):
+	    shutil.move('combined_calls.vcf', '../combined_calls.vcf')
+	if os.path.isfile('annotated.{}_multianno.vcf'.format(ANNOVAR_VERSION)):
+	    shutil.move('annotated.{}_multianno.vcf'.format(ANNOVAR_VERSION),
+	                '../annotated.{}_multianno.vcf'.format(ANNOVAR_VERSION))
+	if os.path.isfile('Tumor_{}_hla_genotype_result.tsv'.format(SAMPLEID)):
+	    shutil.move('Tumor_{}_hla_genotype_result.tsv'.format(SAMPLEID),
+	                '../Tumor_hla_genotype.tsv')
+	if os.path.isfile('Normal_{}_hla_genotype_result.tsv'.format(SAMPLEID)):
+	    shutil.move('Normal_{}_hla_genotype_result.tsv'.format(SAMPLEID),
+	                '../Normal_hla_genotype.tsv')
+	if os.path.isfile('sample1_final.bam'):
+            shutil.move('sample1_final.bam', '../tumor_final.bam')
+	if os.path.isfile('sample2_final.bam'):
+            shutil.move('sample2_final.bam', '../normal_final.bam')
         if os.path.isdir('../{}_bamQCNormal'.format(SAMPLEID)):
             shutil.rmtree(os.path.abspath('../{}_bamQCNormal'.format(SAMPLEID)))
-        shutil.move('bamQC_Normal', '../{}_bamQCNormal'.format(SAMPLEID))
+	if os.path.isdir('bamQC_Normal'):
+            shutil.move('bamQC_Normal', '../{}_bamQCNormal'.format(SAMPLEID))
         if os.path.isdir('../{}_bamQCTumor'.format(SAMPLEID)):
             shutil.rmtree(os.path.abspath('../{}_bamQCTumor'.format(SAMPLEID)))
-        shutil.move('bamQC_Tumor', '../{}_bamQCTumor'.format(SAMPLEID))
+	if os.path.isdir('bamQC_Tumor'):
+            shutil.move('bamQC_Tumor', '../{}_bamQCTumor'.format(SAMPLEID))
         for file in glob.glob('*_fastqc*'):
             shutil.move(file, '../{}_{}'.format(SAMPLEID, file))
         for file in glob.glob('*_trimming_report*'):
